[ui] Move views to a separate folder
Moves the Dashboard and Login views to the
'views' folder.
